requirements.txt
asgiref==3.5.0
Django==4.0.3
django-cors-headers==3.11.0
django-ninja==0.17.0
Pillow==9.1.0
pydantic==1.9.0
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.1
